AURORA-MINDS-Web-app/aurora-minds-FastAPI/requirements.txt
Faker~=25.0.0
fastapi~=0.111.0
pydantic~=2.7.1
pytest~=8.2.0
SQLAlchemy~=2.0.29
starlette~=0.37.2
uvicorn~=0.29.0
python-dotenv~=1.0.1
asyncpg ~=0.29.0
python-multipart ~=0.0.9
pydantic-settings ~=2.2.1
python-jose ~= 3.3.0
bcrypt ~=4.1.2
pytest ~=8.2.0
pytest-asyncio ~=0.23.6
httpx ~=0.27.0